+++ packages/uui-input/lib/uui-input.element.ts
@@ -2,6 +2,7 @@ import { FormControlMixin } from '@umbraco-ui/uui-base/lib/mixins';
 import { defineElement } from '@umbraco-ui/uui-base/lib/registration';
 import { css, html, LitElement, PropertyValueMap } from 'lit';
 import { property, query } from 'lit/decorators.js';
+import { ifDefined } from 'lit/directives/if-defined.js';
 
 import { UUIInputEvent } from './UUIInputEvent';
 
@@ -206,6 +207,15 @@ export class UUIInputElement extends FormControlMixin(LitElement) {
   placeholder = '';
 
   /**
+   * Defines the input autocomplete.
+   * @type {string}
+   * @attr
+   * @default undefined
+   */
+  @property()
+  autocomplete?: string;
+
+  /**
    * This property specifies the type of input that will be rendered.
    * @type {'text' | 'tel'| 'url'| 'email'| 'password'| 'date'| 'month'| 'week'| 'time'| 'datetime-local'| 'number'| 'color'}
    * @attr
@@ -264,13 +274,13 @@ export class UUIInputElement extends FormControlMixin(LitElement) {
     return this._input;
   }
 
-  private _onInput(e: Event) {
+  protected onInput(e: Event) {
     this.value = (e.target as HTMLInputElement).value;
 
-    // TODO: Do we miss an input event?
+    this.dispatchEvent(new UUIInputEvent(UUIInputEvent.INPUT));
   }
 
-  private _onChange() {
+  protected onChange() {
     this.pristine = false;
     this.dispatchEvent(new UUIInputEvent(UUIInputEvent.CHANGE));
   }
@@ -291,13 +301,14 @@ export class UUIInputElement extends FormControlMixin(LitElement) {
         .type=${this.type}
         .value=${this.value as string}
         .name=${this.name}
+        autocomplete=${ifDefined(this.autocomplete as any)}
         placeholder=${this.placeholder}
         aria-label=${this.label}
         .disabled=${this.disabled}
         ?required=${this.required}
         ?readonly=${this.readonly}
-        @input=${this._onInput}
-        @change=${this._onChange} />
+        @input=${this.onInput}
+        @change=${this.onChange} />
       ${this.renderAppend()}
     `;
   }